remedies-bot-master/requirements.txt
requests==2.31.0
google-auth==2.23.3
google-cloud-aiplatform==1.35.0
vertexai==0.0.1
fastapi==0.104.1
pydantic==2.4.2
uvicorn==0.23.2